+++ inst/doc/doges-terms.R
@@ -22,27 +22,27 @@ knitr::kable(doges.one.year[order(doges.one.year$years),],col.names=NULL)
 knitr::kable(doges.all.years[which.max(doges.all.years$years),])
 
 ## ----using.doges.years--------------------------------------------------------
-doges.years.all <- doges.years()
-knitr::kable(head(doges.years.all[order(-doges.years.all$Years),],10))
+data("doges.years")
+knitr::kable(head(doges.years[order(-doges.years$Years),],10))
 
 ## ----distribution, fig.cap="This histogram reproduces, with a fixed bin size, Figure 1 in [@smith2021long]."----
 library(ggplot2)
-ggplot(doges.years.all, aes(x=Years))+geom_histogram(binwidth=5)
+ggplot(doges.years, aes(x=Years))+geom_histogram(binwidth=5)
 
 ## ----boxplot------------------------------------------------------------------
-doges.years.all$Century <- as.factor(doges.years.all$Century)
-ggplot(doges.years.all, aes(x=Century,y=Years))+geom_boxplot()
+doges.years$Century <- as.factor(doges.years$Century)
+ggplot(doges.years, aes(x=Century,y=Years))+geom_boxplot()
 
 ## ----serrata------------------------------------------------------------------
-doges.years.all$pre.serrata <- TRUE
-doges.years.all[doges.years.all$Start >= 1297,]$pre.serrata <- FALSE
-means <- aggregate(Years ~  pre.serrata, doges.years.all, mean)
-medians <- aggregate(Years ~  pre.serrata, doges.years.all, median)
-ggplot(doges.years.all, aes(x=pre.serrata, y=Years))+geom_boxplot(notch=T)+
+doges.years$pre.serrata <- TRUE
+doges.years[doges.years$Start >= 1297,]$pre.serrata <- FALSE
+means <- aggregate(Years ~  pre.serrata, doges.years, mean)
+medians <- aggregate(Years ~  pre.serrata, doges.years, median)
+ggplot(doges.years, aes(x=pre.serrata, y=Years))+geom_boxplot(notch=T)+
   stat_summary(fun=mean, geom="point", shape=20, size=3, color="red", fill="red") +
   geom_text(data = means, aes(label = round(Years, 2), y = Years + 2), size = 3) + 
   geom_text(data = medians, aes(label = round(Years, 2), y = Years - 1), size = 3)
 
 ## ----wilcox-------------------------------------------------------------------
-wilcox.test(doges.years.all[doges.years.all$pre.serrata == T,]$Years,   doges.years.all[doges.years.all$pre.serrata == F,]$Years )
+wilcox.test(doges.years[doges.years$pre.serrata == T,]$Years,   doges.years[doges.years$pre.serrata == F,]$Years )